Capstone/requirements.txt
numpy
transformers
datasets
tqdm
matplotlib
wandb
torchmetrics
torchinfo
pillow
pytorch-lightning
peft
bitsandbytes
einops